First webpage data.
Added the index.html with just an iframe of the actual page, a few css and a title... basic stuff.

Added /CSS/main.css removing margin and padding from body and importing "/CSS/iframe.css" that for now just adjust the iframe size.

The next update will gonna be with more significant changes.